wall compilation flag + warning fixes

+++ src/ecs/Constants.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-constexpr size_t WINDOW_WIDTH = 1200;
-constexpr size_t WINDOW_HEIGHT = 800;
+constexpr int WINDOW_WIDTH = 1200;
+constexpr int WINDOW_HEIGHT = 800;
 
-constexpr size_t TILE_SIZE = 40;
+constexpr int TILE_SIZE = 40;
 constexpr size_t FRAMERATE_LIMIT = 60;
